requirements.txt
numpy
scipy
keras
tensorflow
pillow
imagehash